pyproject.toml
[tool.poetry]
name = "cluster"
version = "0.1.0"
description = "tools"
authors = ["ifechukwudaniel.eth <mailidanielifechukwu@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0.0"
pandas = ">=3.0.2,<4.0.0"
pyarrow = ">=23.0.1,<24.0.0"
psycopg2-binary = ">=2.9.11,<3.0.0"
python-dotenv = ">=1.0.0"
sqlalchemy = ">=2.0.48,<3.0.0"
numpy = ">=2.4.4,<3.0.0"
pydantic = ">=2.12.5,<3.0.0"
sqlmodel = ">=0.0.38,<0.0.39"
langchain-openrouter = ">=0.2.1"
langchain = "^1.2.15"
langgraph = "^1.1.6"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"